LayoutTests/accessibility/select-menulist-crash.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<link rel="stylesheet" href="../fast/js/resources/js-test-style.css">
<script src="../fast/js/resources/js-test-pre.js"></script>
<script>
var successfullyParsed = false;
</script>
</head>
<body>
<form>
<select id="target">
<option selected>foo</option>
<option>bar</option>
</select>
</form>
<p id="description"></p>
<div id="console"></div>
<script>
description('This test is to ensure that we do not crash when selecting an item from a menu list in a list box through the accessibility API.')

if (window.layoutTestController) {
  layoutTestController.dumpAsText();

  if (window.accessibilityController) {
    document.getElementById("target").focus();
    var selectElement = accessibilityController.focusedElement;
    var menuList = selectElement.childAtIndex(0);
    var item1 = menuList.childAtIndex(0);
    var item2 = menuList.childAtIndex(1);

    // Check roles.
    shouldBe("selectElement.role", "'AXRole: combo box'");
    shouldBe("menuList.role", "'AXRole: menu'");
    shouldBe("item1.role", "'AXRole: menu item'");
    shouldBe("item2.role", "'AXRole: menu item'");

    // Execute the press() method for every item in the menu list.
    item1.press();
    item2.press()

    testPassed("Did not crash");
  }
}

successfullyParsed = true;
</script>
<script src="../fast/js/resources/js-test-post.js"></script>
</body>
</html>
